+++ yazelix_default.nix
@@ -47,7 +47,7 @@
 
   # Cursor trail preset
   # Supported by all terminal emulators: "none"
-  # Supported by Ghostty: "blaze", "snow", "cosmic", "ocean", "forest", "sunset", "neon", "party"
+  # Supported by Ghostty: "blaze", "snow", "cosmic", "ocean", "forest", "sunset", "neon", "party", "prism", "orchid"
   # Supported by Ghostty and Kitty: "snow"
   # In Short: only ghostty supports all the cool cursor trails
   cursor_trail = "blaze";